--- a/schemas/db_schema.go
+++ b/schemas/db_schema.go
@@ -1,12 +1,15 @@
 package schemas
 
+import "sync"
+
 type DbStore interface {
 	Setup() error
 	UseDatabase(name string) error
 	GetTableFields(database, table string) ([]TableFields, error)
 	GenerateInsertionMap(fields []TableFields, seedSize int64) []map[string]InsertionMap
-	BatchInsertFromMap(bArr []map[string]InsertionMap, fields []TableFields, table string, chunkSize int64) error
-	SelectCount(table string) (int64, error)
+	BatchInsertFromMap(bArr []map[string]InsertionMap, fields []TableFields, table string, chunkSize int64, dbName string, maxConn int, wg *sync.WaitGroup) error
+	SelectCount(table string, dbName string) (int64, error)
+	GetMaxConnections() (int, error)
 }
 
 type TableFields struct {
@@ -25,4 +28,8 @@ type InsertionMap struct {
 
 type NumberNil interface {
 	Number() int64
+}
+type ShowVariables struct {
+	Variable_name string `db:"Variable_name"`
+	Value         string `db:"Value"`
 }