# Changes:

### Changes to BatchInsertFromMap:

- Changed algorithm to INSERT batch when array meets chunkSize, instead of saving to slice.
- Added GO routines to INSERT lock the SQL string builder generation
- Added Wait Group to wait for all INSERTS to conclude
- Added Channels to limit the number of GO routines for the configured database max connections
- Removed Printf calls from the inside of loops.

### Lazy Benchmarks
- 1M rows, 100 chunkSize: 39.4s -> 9.7s
- 1M rows, 1k chunkSize:22.8s -> 8.3s
- 1M rows, 10k chunkSize: 19.3s -> 8.2s
- 1M rows, 100k chunkSize: 19.6s -> 15.6s

The bigger the difference from the numbers of rows inserted compared to the chunkSize, the bigger the difference on performance with this change.